+++ corda-testacles-common/src/main/kotlin/com/github/manosbatsis/corda/testacles/common/MapToInstanceHelper.kt
@@ -1,0 +1,7 @@
+package com.github.manosbatsis.corda.testacles.common
+
+class MapToInstanceHelper {
+    companion object{
+
+    }
+}