+++ corda-testacles-model-api/src/main/kotlin/com/github/manosbatsis/corda/testacles/model/api/MapToInstanceHelper.kt
@@ -1,7 +1,0 @@
-package com.github.manosbatsis.corda.testacles.model.api
-
-class MapToInstanceHelper {
-    companion object{
-
-    }
-}