+++ corda-testacles-processor/src/main/kotlin/com/github/manosbatsis/corda/testacles/processor/support/LiteDtoMemberStrategy.kt
@@ -22,7 +22,7 @@
 package com.github.manosbatsis.corda.testacles.processor.support
 
 import co.paralleluniverse.fibers.Suspendable
-import com.github.manosbatsis.corda.testacles.model.api.ModelAdapter
+import com.github.manosbatsis.corda.testacles.common.ModelAdapter
 import com.github.manotbatsis.kotlin.utils.kapt.dto.strategy.DtoMembersStrategy.Statement
 import com.github.manotbatsis.kotlin.utils.kapt.dto.strategy.DtoNameStrategy
 import com.github.manotbatsis.kotlin.utils.kapt.dto.strategy.SimpleDtoMembersStrategy